OY-3894 : PR-huomioiden korjauksia: Resulmodal omaksi komponentiksi
