+++ src/main/app/.env
@@ -6,7 +6,7 @@ REACT_APP_CALLER_ID=1.2.246.562.10.00000000001.koutaui
 
 REACT_APP_DEV_SERVER_URL=https://localhost:3000
 
-REACT_APP_KONFO_URL=https://hahtuvaopintopolku.fi/konfo
+REACT_APP_KONFO_URL=https://untuvaopintopolku.fi/konfo
 
 # All requests except / and /kouta/* are proxied to here when running 'npm start'
 DEV_VIRKAILIJA_URL=https://virkailija.hahtuvaopintopolku.fi
