+++ src/main/app/src/pages/HomePage/CopyResultModal.tsx
@@ -1,18 +1,14 @@
-import React, { useCallback, useMemo } from 'react';
+import React, { useCallback } from 'react';
 
 import _fp from 'lodash/fp';
-import { useTranslation } from 'react-i18next';
 import { UseMutationResult } from 'react-query';
-import { usePrevious } from 'react-use';
 import styled from 'styled-components';
 
 import { RouterAnchor } from '#/src/components/Anchor';
-import ErrorAlert from '#/src/components/ErrorAlert';
-import ListTable from '#/src/components/ListTable';
-import Modal from '#/src/components/Modal';
-import { Box, Button, Icon } from '#/src/components/virkailija';
+import { Icon } from '#/src/components/virkailija';
 import { ENTITY } from '#/src/constants';
 import { useUserLanguage } from '#/src/hooks/useUserLanguage';
+import { ResultModal } from '#/src/pages/HomePage/ResultModal';
 import { getFirstLanguageValue } from '#/src/utils/languageUtils';
 
 import { useEntitySelection } from './useEntitySelection';
@@ -71,28 +67,12 @@ const useTableColumns = (t, entityType, getLinkUrl) => [
   },
 ];
 
-export const CopyResultList = ({ data, entityType, getLinkUrl }) => {
-  const { t } = useTranslation();
-  const columns = useTableColumns(t, entityType, getLinkUrl);
-  const rows = useMemo(
-    () => _fp.map(result => ({ ...result, key: result.oid }), data),
-    [data]
-  );
-
-  return <ListTable rows={rows} columns={columns} />;
-};
-
 type CopyResultItem = {
   oid: string;
   created: {
     toteutusOid?: string;
     hakukohdeOid?: string;
   };
-};
-
-const usePreviousNonNil = value => {
-  const prev = usePrevious(value);
-  return _fp.isEmpty(value) ? prev : value;
 };
 
 const isCopyResultSuccessful = mutationResult =>
@@ -114,10 +94,6 @@ export const CopyResultModal = ({
   >;
   getLinkUrl: any;
 }) => {
-  const { t } = useTranslation();
-
-  const isOpen = ['success', 'error'].includes(mutationResult.status);
-
   const { removeSelection } = useEntitySelection(entityType);
 
   const onClose = useCallback(() => {
@@ -127,32 +103,14 @@ export const CopyResultModal = ({
     mutationResult.reset();
   }, [mutationResult, removeSelection]);
 
-  const data = usePreviousNonNil(mutationResult.data);
-
   return (
-    <Modal
-      minHeight="90vh"
-      maxWidth="1200px"
-      open={isOpen}
+    <ResultModal
       onClose={onClose}
-      header={headerText}
-      footer={
-        <Box display="flex" justifyContent="flex-end">
-          <Button variant="outlined" onClick={onClose}>
-            {t('yleiset.sulje')}
-          </Button>
-        </Box>
-      }
-    >
-      {mutationResult.isError ? (
-        <ErrorAlert center>{t('etusivu.kopiointiEpaonnistui')}</ErrorAlert>
-      ) : (
-        <CopyResultList
-          data={data}
-          entityType={entityType}
-          getLinkUrl={getLinkUrl}
-        />
-      )}
-    </Modal>
+      headerText={headerText}
+      mutationResult={mutationResult}
+      entityType={entityType}
+      getLinkUrl={getLinkUrl}
+      useTableColumns={useTableColumns}
+    />
   );
 };
